Add a new event - backend.auth.afterLogin
I wanted to listen to an event when a users logs-in and I realised that October does not yet have that. It would be great to support this out of the box to avoid users make changes to core files or listen to hack-ish events to do something after login.