Tighten up the get/set event logic
The internal events beforeGetAttribute, beforeSetAttribute, getAttribute, setAttribute now apply strictly to local attributes, not relations. This should improvement performance and consistency.

The attributesToArray method (used for casting model to an array) has been brought in to line with the object behavior:
- Jsonable attributes are applied (previously string was returned)
- The internal events are fired for each attribute

Refs https://github.com/rainlab/translate-plugin/issues/188

Noticed a collision with the "getRelationValue" method, it behaves differently in Laravel than in October. This may need to be renamed in future if it causes issues with third party libraries. Another consideration is if an event is needed for "get relation", this logic will need to be moved from getAttribute to getRelationValue.