Progress on database upgrade
There was an attempt made to retain array over collection when returning results, however, too many internals are now expecting a collection so we have to roll over on this one. Fortunately the lists() method was removed entirely--one of the primary differentiators in Laravel 5.1 that we retained as returning an array--so we were able to make this method our own and keep it returning a basic array. Now the only difference is ->get() will return a collection instead of an array. We'll need a release note about this.

Some of the internals of the relationship objects have been "cleaned up". Still deciding if we should retain the old naming convention, essentially duplicating properties, or make an attempt to shift to the newer, more verbose convention. TBA

Still unsure the exact nature of new scopes, perhaps they have been upgraded properly, although it will need testing.

Blueprint no longer requires any overrides--fixes not present in 5.1 are now found in 5.5--and remains as a stub class.