Explicitly provide an argument separator
This fixes an issue where the default PHP setting has the arg_separator.output set to &amp; but this setting is commented out, leaving the system default setting as plain & character. It is a mismatch in default configuration and system default configuration that causes updates to fail for seemingly mysterious reasons.

Refs https://twitter.com/DarkCisum/status/916087715631951872