Merge dynamic properties to the base attributes when hydrating halcyon (#366)
This merges the dynamically added properties to the base attributes when hydrating Halcyon models to persist the values of those properties. The addPurgeable() method was also added and dynamically added properties are now automatically purged along with purgeable properties bringing the behavior more in line with Eloquent models. This is related to https://github.com/rainlab/translate-plugin/pull/367