Retrieve the hasOne simple value from the local key, not the foreign key
HasOne's getSimpleValue was erroneously using the foreign key of the relationship to retrieve the simple value from the related record, this resulted in only ever returning the primary key of the owning / parent record instead of the actual key for the related record. Switched it to use the localKey (defaults to id) instead to retrieve the simple value of the relationship.