Change POSTFIELDS back to array
In a previous commit, this has been changed to a string to ensure the correct argument separator to be set. However, starting from PHP 5.2.0, this disables us from uploading files, as for files an array argument is required. From the docs:
``If value is an array, the Content-Type header will be set to multipart/form-data. As of PHP 5.2.0, value must be an array if files are passed to this option with the @ prefix. As of PHP 5.5.0, the @ prefix is deprecated and files can be sent using CURLFile.
``