Make dynamic subjects available to mail templates
Accomplishes this by firing the message callback before content is added to the message.

This fixes an issue where subjects set on the message object through the callback were not able to override a mail template's subject as they're supposed to in System\Classes\MailManager because the callback was called after the MailManager's listener for mailer.beforeAddContent had already fired. The original issue was reported and fixed in octobercms/october#1665. However, that fix was then later broken by https://github.com/octobercms/library/commit/f409ad425e68e38adf8917f80769ea6af828f267 which was for the L5.5 upgrade in Build 420. It's also worth noting that Laravel has moved the callback to be before the content is added starting in L5.6, so this change also brings us closer in line with Laravel for the future 6.0 upgrade (see https://github.com/laravel/framework/pull/22995 for that commit and background information).