Add Laravel sig support to simplePaginate
Also did a consistency clean up, args are ordered as close to Laravel as possible and are now consistent with paginate(). This may cause some pain for people using columns and pageName on simplePaginate by swapping the last two args, which somehow slipped through the review process, hopefully rare and is why nobody noticed. In short: The simplePaginate method args now match the paginate method args

Refs https://github.com/octobercms/october/issues/4849