Inject queue name into auto-created Mailables. (#454)
The docs stipulate using a string for the view parameter when using Mail. When a string is given, the Mailer class is converting the string into a Mailable object to pass through Laravel's mail functionality. This Mailable object is not being populated with the queue name, and is instead trying to use the queue name as the manager that is passed through to the Mailable object. This is failing when the Mailable is added to the queue, therefore breaking the Mail::queueOn() and Mail::laterOn() functions.

This fix takes the queue name and injects it into the Mailable object, then sets the queue parameter to null so that the default QueueManager is used to create the queued job.

Fixes https://github.com/octobercms/october/issues/4870.