Add e() helper into October helpers
In Laravel 5.6, they changed $doubleEncode to "true" by default. As a LOT of functionality in OCMS has no $doubleEncode specified, so this is resulting in a lot of double-encoded content (ie. content for code editors or rich editors).

Our override brings the default back to "false" to allow all current functionality to work without change.