Guess extension for uploaded files if missing (#452)
This change aims to guess file extension from MIME type if file is being uploaded without file extension. The guessed extension is then added to the generated filename.

Some contextual explanation: 

Why would a file be uploaded without extension?

It's the case for raw blob files. For instance, in my app I let the user choose a image file and then allow her/him to do some basic editing (like cropping) client-side. Thus, when uploading, I'm no longer uploading the raw file, but a blob file that don't have a filename and thus doesn't have any extension.

Why storing file without extension is an issue?

This can break other October features. For instance, the issue I had when uploading blob images, was that the getThumb function silently failed (because it didn't consider images without extension as images, and thus do not resize them). So for a long time, I was serving high-res image on my website just to display as thumbnail and didn't notice it.

You could argue that a solution is to change my client-side code to include a filename and manually add an extension. The thing is, even on a regular form input field, a user may attempt to upload an image without file extension, that successfully passes the "image" validation rule (because this rule only check MIME type), but won't work with the getThumb function.