Allow deep symlinks when translating paths to local URLs
This extends the work done by @LukeTowers in https://github.com/octobercms/library/commit/a88b5fd3880edeac0c2bb3bf9e5297c508300055 by doing a search (once) for all folders within the base directory and determining symlinks within.

It will create an array of sources and targets in order to translate these to URLs.

In addition, this will respect the cms.restrictBaseDir config value, in that symlinks that link to folders outside the base directory will not be recorded if restricted to the base directory, but symlinks that remain within the base directory will still work.