Add config value to control loading of discovered Laravel packages.
Adds the config value 'app.loadDiscoveredPackages' (default: true) - if true, packages discovered through Laravel's autodiscovery feature will be loaded. The reason for this control is that in some cases, packages which are autodiscovered belong to plugins. When the plugins themselves are disabled, the discovered package is still being loaded. See https://github.com/rainlab/debugbar-plugin/issues/34 for an example.