Support session keys passed through createMany() (#466)
When calling $model->someBelongsToManyRelation()->createMany([...]) a deferred binding is created in the database with session key 0. Calling $model->save() after this does not commit the deferred binding. A deferred binding is created even when $model exists in the database.

createMany() is called through the Illuminate file vendor/laravel/framework/src/Illuminate/Database/Eloquent/Relations/BelongsToMany.php.
It loops through each passed array calling $this->create($record, (array) ($joinings[$key] ?? []), false); - notice that third argument passed (the one for sessionKey) is false.

create() is our own method called in vendor/october/rain/src/Database/Relations/BelongsToMany.php.
It calls $this->add($model, $sessionKey, $pivotData); which in turn checks for a $sessionKey of null, and if not found, calls $this->parent->bindDeferred($this->relationName, $model, $sessionKey);. Because our sessionKey is false instead of null this is where we end up.

This may require revisiting in the future to support the third argument Laravel is passing ($touch) properly, but this looks fine for now.