Attempt to handle failures when loading packages/services from cache.
In certain cases (for example, with Composer), sites will break when a package removed by Composer is loaded. This is because discovered packages and services are stored in cache ("storage/framework/services.php" and "storage/framework/packages.php").

This change will attempt to use the normal mechanism to load packages and services, and if it fails for any reason, it will delete the package and service cache and attempt to re-load the packages and services one more time (emulating someone running "php artisan config:clear" before "php artisan package:discover"). If a failure still occurs, then the failure is made known to the user.

Refs: https://github.com/octobercms/library/pull/492#issuecomment-636434950