Add config value to control loading of discovered Laravel packages (#492)
Adds the config value 'app.loadDiscoveredPackages' (default: false) - if true, packages discovered through Laravel's autodiscovery feature will be loaded. The reason for this control is that in some cases, packages which are autodiscovered belong to plugins. When the plugins themselves are disabled, the discovered package is still being loaded. See https://github.com/rainlab/debugbar-plugin/issues/34 for an example.

Additionally, in certain cases (for example, with Composer), sites will break when a package removed by Composer is loaded. This is because discovered packages and services are stored in cache ("storage/framework/services.php" and "storage/framework/packages.php").

This change will attempt to use the normal mechanism to load packages and services, and if it fails for any reason, it will delete the package and service cache and attempt to re-load the packages and services one more time (emulating someone running "php artisan config:clear" before "php artisan package:discover"). If a failure still occurs, then the failure is made known to the user.

Refs: https://github.com/octobercms/library/pull/492#issuecomment-636434950