Fix one char discrepancy between limited HTML strings and break types
Windows CRLF line breaks are considered to be two characters, whereas Unix LF line breaks are one. This causes a discrepancy between Windows and Linux for strings that use HtmlBuilder::limit(), where HTML strings that are limited in Windows will be shorter than on Linux.

This adds a new method to HtmlBuilder which will get the "readable" length of a string and consider both line break types to one character, as well as any extra spacing in between characters which is rendered as a single space in HTML.

Linked to https://github.com/octobercms/library/pull/510.