allowTrashedSlugs is redundant
This is a complicated way to solve the issue of applied global scopes. Slugs are unique to the record regardless of their soft deletion state and should not be concerned with scopes at all. In the unusual case where a slug may need to be shared across multiple records, the solution is to override the trait logic (newSluggableQuery) to implement the custom logic, although it is doubtful that this is even a thing and therefore predictably a non-breaking change