Adds internal marker for forced validation
This fixes an issue in the original design where the before/after events will always fire, even during force, since they can be used for manipulating values prior to validation where beforeSave occurs too late -- this is in addition to performing actual validation tasks. In these cases the custom validation logic is oblivious to whether the event was forced or not. The marker allows a simple check skip method.

```
public function beforeValidate()
{
    if ($this->validationForced) {
        return;
    }

    // ...
}
```