Rollback 4914945cbf20e12a414549df40440320789630fb
This was suggested by a linter but will be problematic later on (see 58efb6f419f2d34f83b866ccd1a81380251ade90). The Laravel favoured ValidationException class does not pass a "previous" so the linter rule needs to be corrected (non-reportable/userland exception). The potential problem will occur when the Validator passes more arguments to the ValidationException, and we need to reserve space for this compatibility requirement.